会の追加周りのコンポーネントを整理
